Add rules command to Discord bot (#136)
Either the full list of rules can be summoned, or a specific one. Input
is validated by Discord.

The rules are defined in a file under the constants folder. If wanted,
we could update this to pull from the message in the rules channel, but
then the message would need to follow a predefined pattern or things
would break.

The bot template included a ratelimiting feature which I thought would
be useful to prevent people from spamming large chunks of text. You can
see it sort of working in the demo below. Not sure why it let some
requests through that were clearly more than 2 per 30 seconds.


[Screencast_20250713_230319.webm](https://github.com/user-attachments/assets/97bd9177-6258-4a25-8946-8b24aad1547d)

Closes #118