Added webhooks to Pragmatic Papers (#153)
notes: as of now, the key `fetch` call is aimed at the [discord
api](https://discord.com/developers/docs/resources/webhook#execute-webhook).
avatar URL could probably be moved to a more accessible spot, the check
for whether or not to actually push a webhook event could probably be
more sophisticated. at the moment it's a volume going draft -> published
while having no previous publication date (i.e. the first publish)

this *just* pushes the raw URL, nothing fancy. having a bot grab it and
do fancy stuff is probably the play, as of right now you could "add
webhook" in the discord server and the message would be the URL (with a
username / avatar override)

partially addresses #119 / #76

---------

Co-authored-by: Patrick Orave <42103411+linkion@users.noreply.github.com>